finish service card gallery
