+++ src/assets/serviceGalleryData.js
@@ -1,0 +1,39 @@
+const galleryData = [
+  {
+    id: '1',
+    title: "צ'אט עם נציג",
+    image: require('../assets/B1-assets/elementsServiceIconsChat.png'),
+  },
+  {
+    id: '2',
+    title: 'תקלות בדרך',
+    image: require('../assets/B1-assets/elementsServiceIconsRoadsideAssistance.png'),
+  },
+  {
+    id: '3',
+    title: 'זימון שירותי מוסך',
+    image: require('../assets/B1-assets/elementsServiceIconsServiceCenter.png'),
+  },
+  {
+    id: '4',
+    title: 'אביזרים',
+    image: require('../assets/B1-assets/elementsServiceIconsHardwareService.png'),
+  },
+  {
+    id: '5',
+    title: "צמיגים ופנצ'ריות",
+    image: require('../assets/B1-assets/elementsServiceIconsTire.png'),
+  },
+  {
+    id: '6',
+    title: 'רכב חשמלי',
+    image: require('../assets/B1-assets/elementsServiceIconsEv.png'),
+  },
+  {
+    id: '7',
+    title: 'שטיפת רכב',
+    image: require('../assets/B1-assets/elementsServiceIconsCarWash.png'),
+  },
+];
+
+export default galleryData;
